requirements.txt
requests	
aiogram
python-dotenv
fastapi
uvicorn
